flatpak: install pycodestyle during build
This will allow us to run `make check` during CI flatpak builds, which
manually build the d-feet module rather than building it with
flatpak-builder. We still use --disable-tests in the real manifest
because the tests require an X (or Wayland) server.

Specifying "cleanup": ["*"] causes all files installed by the
pycodestyle module to be removed at the end of the build process.